API for total Sale AMount for Current Day

+++ src/main/java/com/tdd/eCommerceApp/entities/Product.java
@@ -21,7 +21,7 @@ public class Product {
     private String name;
 
 
-    private Integer price;
+    private Double price;
 
     private Integer quantity;
 
@@ -57,7 +57,7 @@ public class Product {
         return price;
     }
 
-    public void setPrice(Integer price) {
+    public void setPrice(double price) {
         this.price = price;
     }
